engine/requirements.txt
tastytrade>=12.0.0
httpx>=0.27.0
aiohttp>=3.9.0
apscheduler>=3.10.4
python-dotenv>=1.0.0
pandas>=2.2.0
numpy>=1.26.0
